Se modifica para emepzar con el crud de productos

+++ Punto_Venta/src/main/java/mx/aragon/unam/repository/producto/ProductoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductoRepository extends JpaRepository<ProductoEntity, Long> {
+public interface ProductoRepository extends JpaRepository<ProductoEntity, Integer> {
 }
